apps/backend/services/ai-agent/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.0
langchain-community>=0.2.0
langfuse>=2.30.0
httpx>=0.27.0
pydantic-settings>=2.2.0
supabase>=2.3.0
